touroute/requirements.txt
Django==1.11.15 
django-common==0.1.51
django-crispy-forms==1.7.2
numpy==1.16.2
oauth2==1.9.0.post1
oauthlib==2.0.6
pandas==0.24.1
pandas-datareader==0.7.0
requests==2.18.4
urllib3==1.22

